+++ ontop-wms/backend/src/main/java/com/ontop/wms/dto/AuthResponse.java
@@ -1,4 +1,4 @@
 package com.ontop.wms.dto;
 
-public record AuthResponse(String token, String username, String role, Integer warehouseId, String warehouseName) {
+public record AuthResponse(String token, String username, String role, Long warehouseId, String warehouseName) {
 }
