+++ ontop-wms/backend/src/main/java/com/ontop/wms/dto/InventoryRequest.java
@@ -12,7 +12,7 @@ public class InventoryRequest {
     private String receiptCode; // for inbound
     private String issueCode;   // for outbound
     private String status;
-    private Integer warehouseId;
+    private Long warehouseId; // Changed to Long
     
     // Circular 200 Fields
     private String delivererName;
