+++ ontop-wms/backend/src/main/java/com/ontop/wms/repository/WarehouseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ontop.wms.entity.Warehouse;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
     Optional<Warehouse> findByCode(String code);
 }
